Move product photos code to separate module
