+++ src/products/product-photos/product-photos.controller.ts
@@ -1,0 +1,107 @@
+import {
+  Controller,
+  Delete,
+  FileTypeValidator,
+  Get,
+  MaxFileSizeValidator,
+  Param,
+  ParseBoolPipe,
+  ParseFilePipe,
+  ParseIntPipe,
+  Post,
+  Query,
+  StreamableFile,
+  UploadedFile,
+  UseInterceptors,
+} from '@nestjs/common';
+import {
+  ApiBody,
+  ApiConsumes,
+  ApiCreatedResponse,
+  ApiForbiddenResponse,
+  ApiNotFoundResponse,
+  ApiOkResponse,
+  ApiProduces,
+  ApiTags,
+  ApiUnauthorizedResponse,
+} from '@nestjs/swagger';
+import { Roles } from '../../auth/decorators/roles.decorator';
+import { Role } from '../../users/entities/role.enum';
+import { Product } from '../entities/product.entity';
+import { FileInterceptor } from '@nestjs/platform-express';
+import { ProductPhotosService } from './product-photos.service';
+
+@ApiTags('products')
+@Controller('products/:id')
+export class ProductPhotosController {
+  constructor(private productPhotosService: ProductPhotosService) {}
+
+  @Get('photos/:photoId')
+  @ApiOkResponse({
+    schema: {
+      type: 'string',
+      format: 'binary',
+    },
+    description: 'Product photo with given id',
+  })
+  @ApiProduces('image/*')
+  @ApiNotFoundResponse({ description: 'Product photo not found' })
+  async getProductPhoto(
+    @Param('id', ParseIntPipe) id: number,
+    @Param('photoId', ParseIntPipe) photoId: number,
+    @Query('thumbnail', ParseBoolPipe) thumbnail?: boolean,
+  ): Promise<StreamableFile> {
+    return await this.productPhotosService.getProductPhoto(
+      id,
+      photoId,
+      thumbnail ?? false,
+    );
+  }
+
+  @Post('photos')
+  @Roles(Role.Admin, Role.Manager)
+  @ApiUnauthorizedResponse({ description: 'User not logged in' })
+  @ApiForbiddenResponse({ description: 'User not authorized' })
+  @ApiNotFoundResponse({ description: 'Product not found' })
+  @ApiCreatedResponse({ type: Product, description: 'Product photo added' })
+  @ApiBody({
+    schema: {
+      type: 'object',
+      properties: {
+        file: {
+          type: 'string',
+          format: 'binary',
+        },
+      },
+    },
+  })
+  @ApiConsumes('multipart/form-data')
+  @UseInterceptors(FileInterceptor('file'))
+  async addProductPhoto(
+    @Param('id', ParseIntPipe) id: number,
+    @UploadedFile(
+      new ParseFilePipe({
+        validators: [
+          new MaxFileSizeValidator({ maxSize: 1024 * 1024 * 10 }),
+          new FileTypeValidator({ fileType: /^image\/(png|jpe?g|gif|webp)/ }),
+        ],
+      }),
+    )
+    file: Express.Multer.File,
+  ): Promise<Product> {
+    return await this.productPhotosService.addProductPhoto(id, file);
+  }
+
+  @Delete('photos/:photoId')
+  @Roles(Role.Admin, Role.Manager)
+  @ApiUnauthorizedResponse({ description: 'User not logged in' })
+  @ApiForbiddenResponse({ description: 'User not authorized' })
+  @ApiNotFoundResponse({ description: 'Product not found' })
+  @ApiOkResponse({ type: Product, description: 'Product photo deleted' })
+  async deleteProductPhoto(
+    @Param('id', ParseIntPipe) id: number,
+    @Param('photoId', ParseIntPipe) photoId: number,
+  ): Promise<Product> {
+    return await this.productPhotosService.deleteProductPhoto(id, photoId);
+  }
+}
