+++ src/products/product-photos/product-photos.service.ts
@@ -1,0 +1,64 @@
+import { Injectable, StreamableFile } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { ProductPhoto } from '../entities/product-photo.entity';
+import { Repository } from 'typeorm';
+import { LocalFilesService } from '../../local-files/local-files.service';
+import { NotFoundError } from '../../errors/not-found.error';
+import { Product } from '../entities/product.entity';
+
+@Injectable()
+export class ProductPhotosService {
+  constructor(
+    @InjectRepository(Product) private productsRepository: Repository<Product>,
+    @InjectRepository(ProductPhoto)
+    private productPhotosRepository: Repository<ProductPhoto>,
+    private localFilesService: LocalFilesService,
+  ) {}
+
+  async getProductPhoto(
+    productId: number,
+    photoId: number,
+    thumbnail: boolean,
+  ): Promise<StreamableFile> {
+    const productPhoto = await this.productPhotosRepository.findOne({
+      where: { id: photoId, product: { id: productId } },
+    });
+    if (!productPhoto) {
+      throw new NotFoundError('product photo', 'id', photoId.toString());
+    }
+
+    const filepath = thumbnail ? productPhoto.thumbnailPath : productPhoto.path;
+
+    const mimeType = thumbnail ? 'image/jpeg' : productPhoto.mimeType;
+
+    return await this.localFilesService.getPhoto(filepath, mimeType);
+  }
+
+  async addProductPhoto(
+    id: number,
+    file: Express.Multer.File,
+  ): Promise<Product> {
+    const product = await this.productsRepository.findOne({ where: { id } });
+    if (!product) {
+      throw new NotFoundError('product', 'id', id.toString());
+    }
+    const photo = new ProductPhoto();
+    const { path, mimeType } = await this.localFilesService.savePhoto(file);
+    photo.path = path;
+    photo.mimeType = mimeType;
+    photo.thumbnailPath = await this.localFilesService.createPhotoThumbnail(
+      file.path,
+    );
+    product.photos.push(photo);
+    return this.productsRepository.save(product);
+  }
+
+  async deleteProductPhoto(id: number, photoId: number): Promise<Product> {
+    const product = await this.productsRepository.findOne({ where: { id } });
+    if (!product) {
+      throw new NotFoundError('product', 'id', id.toString());
+    }
+    product.photos = product.photos.filter((p) => p.id !== photoId);
+    return this.productsRepository.save(product);
+  }
+}
