+++ src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, StreamableFile } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Product } from './entities/product.entity';
 import { In, Repository } from 'typeorm';
@@ -6,7 +6,6 @@ import { ProductCreateDto } from './dto/product-create.dto';
 import { ProductUpdateDto } from './dto/product-update.dto';
 import { Attribute } from './entities/attribute.entity';
 import { AttributeDto } from './dto/attribute.dto';
-import { ProductPhoto } from './entities/product-photo.entity';
 import { NotFoundError } from '../errors/not-found.error';
 import { AttributeType } from './entities/attribute-type.entity';
 import {
@@ -19,8 +18,8 @@ import { TypeCheckError } from '../errors/type-check.error';
 import { parse } from 'json2csv';
 import * as csv from 'csvtojson';
 import { OrderItem } from '../orders/entities/order-item.entity';
-import { LocalFilesService } from '../local-files/local-files.service';
 import { AttributeValueType } from './entities/attribute-value-type.enum';
+import { ProductPhoto } from './entities/product-photo.entity';
 
 @Injectable()
 export class ProductsService {
@@ -30,9 +29,6 @@ export class ProductsService {
     private attributesRepository: Repository<Attribute>,
     @InjectRepository(AttributeType)
     private attributeTypesRepository: Repository<AttributeType>,
-    @InjectRepository(ProductPhoto)
-    private productPhotosRepository: Repository<ProductPhoto>,
-    private localFilesService: LocalFilesService,
   ) {}
 
   async getProducts(): Promise<Product[]> {
@@ -157,53 +153,6 @@ export class ProductsService {
         throw new TypeCheckError('attribute value', check[0]);
       }
     });
-  }
-
-  async getProductPhoto(
-    productId: number,
-    photoId: number,
-    thumbnail: boolean,
-  ): Promise<StreamableFile> {
-    const productPhoto = await this.productPhotosRepository.findOne({
-      where: { id: photoId, product: { id: productId } },
-    });
-    if (!productPhoto) {
-      throw new NotFoundError('product photo', 'id', photoId.toString());
-    }
-
-    const filepath = thumbnail ? productPhoto.thumbnailPath : productPhoto.path;
-
-    const mimeType = thumbnail ? 'image/jpeg' : productPhoto.mimeType;
-
-    return await this.localFilesService.getPhoto(filepath, mimeType);
-  }
-
-  async addProductPhoto(
-    id: number,
-    file: Express.Multer.File,
-  ): Promise<Product> {
-    const product = await this.productsRepository.findOne({ where: { id } });
-    if (!product) {
-      throw new NotFoundError('product', 'id', id.toString());
-    }
-    const photo = new ProductPhoto();
-    const { path, mimeType } = await this.localFilesService.savePhoto(file);
-    photo.path = path;
-    photo.mimeType = mimeType;
-    photo.thumbnailPath = await this.localFilesService.createPhotoThumbnail(
-      file.path,
-    );
-    product.photos.push(photo);
-    return this.productsRepository.save(product);
-  }
-
-  async deleteProductPhoto(id: number, photoId: number): Promise<Product> {
-    const product = await this.productsRepository.findOne({ where: { id } });
-    if (!product) {
-      throw new NotFoundError('product', 'id', id.toString());
-    }
-    product.photos = product.photos.filter((p) => p.id !== photoId);
-    return this.productsRepository.save(product);
   }
 
   async exportProducts(): Promise<string> {
